more safety checks and features
currently it's more of a rearrangment of the original code, with some
added safety/integrity checks
`resume` is now a separate function, the plan is to have the code for
actually resuming within this function and code for locating the device
outside.
in the more general case, if `late_resume` is enabled, the initram will
check at both the `init_early` and again at the `init_premount` stage
for the presence of the resume device. Once host-only is implemented
then it'll be able to determine at build time which state it'll need to
check at.
For safety checks, currently it checks whether the device is a
valid block device, whether resume has already been attempted previous
(based on the content of `/sys/power/resume`) & whether there are are
any other mounted block devices.