improve error messages/exceptions for failed header validation
LUKS1 is only supported as much as it needs to be, don't validate LUKS1
headers and provide info about configuring for detached headers or
headers which should not be validated.

Raise an autodetect error if the header cannot be found and a header
file is not specified, validation error if a header file is specified
and cannot be validated.

Signed-off-by: Zen <z@pyl.onl>