add _late_args for later kernel_version
in cases where --kver is passed by cmdline (such as when used with
installkernel), validation could fail if there is no modules directory
before the config.toml is read where no_kmod may be set.

this allows for certain config to only be processed in the later stages
so cmdline args can be applied in the final config pass.

Signed-off-by: Zen <z@pyl.onl>